feat: Update Product and Purchase components with async select and improved UI
- Changed default sort order in Product component from ascending to descending.
- Refactored ProductForm to use AsyncSelect for category and combination fields, allowing for dynamic loading of options.
- Updated form data structure to handle numeric fields more effectively.
- Enhanced UI with gradient backgrounds and improved layout for better user experience.
- Added new endpoints for fetching combinations and categories.
- Updated PurchaseInvoiceForm to improve layout and user feedback on submission.
- Refactored commonDataTable to improve checkbox handling and button states based on order status.
- Increased pagination limits in sales order and sales indent constants for better data handling.

+++ Warehouse_Inward/frontend/src/components/GoodReceiptNote/grn.jsx
@@ -11,6 +11,7 @@ import {
 } from "../../constant/grnConstant.js";
 import { ROUTES } from "../../constant/routePath.js";
 import { LIMITPAGE } from "../../constant/grnConstant.js";
+import { STATUS } from "../../constant/constant.js";
 
 function GRNList() {
   const [page, setPage] = useState(1);
@@ -46,7 +47,12 @@ function GRNList() {
     ALLEndpoint.GRNEndpoints.deleteGRN.method
   );
 
-  const handleDelete = (grn_id) => {
+  const handleDelete = (grn) => {
+    if([STATUS.COMPLETED,STATUS.CANCELLED].includes(grn.status)){
+      alert(`Cannot delete a ${grn.status} GRN.`);
+      return;
+    }
+    const grn_id = grn_id.id;
     deleteItem({
       idField: "grn_id",
       idValue: grn_id,
@@ -91,15 +97,20 @@ function GRNList() {
       onView={(grn) =>
         navigate(ROUTES.GRN.VIEW('grn',grn.id))
       }
-      onDelete={(grn) => handleDelete(grn.id)}
-      extraAction={(grn) => (
+      onDelete={(grn) => handleDelete(grn)}
+      extraAction={(grn) => { 
+        const isDisabled = [STATUS.COMPLETED,STATUS.CANCELLED].includes(grn.status);
+
+        return(
         <button
+          disabled={isDisabled}
           onClick={() => navigate(ROUTES.PURCHASE_INVOICE.ADD(grn.id))}
-          className="bg-blue-500 text-white px-3 py-1 rounded-md hover:bg-blue-600 transition-colors"
+          className={`bg-blue-500 text-white px-3 py-1 rounded-md hover:bg-blue-600 transition-colors ${isDisabled ? "opacity-50 cursor-not-allowed" : ""
+              }}`}
         >
           Create Invoice
         </button>
-      )}
+      )}}
     />
   );
 }
